return int

+++ python_variable_annotations/2-floor.py
@@ -2,6 +2,6 @@
 """Module contains function floor"""
 
 
-def find_floor(float: n) -> float:
+def find_floor(float: n) -> int:
     """Function that returns the floor of a float"""
     return floor(n)
